refactor: 検索関連のファイルをfeaturesディレクトリに集約
