+++ src/features/search/components/SearchBar/SearchBar.module.scss
