+++ src/features/search/components/SearchResultClient/SearchResultClient.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import PostList from '@/features/posts/components/PostList/PostList';
 import { TimelineContext, type TimelineContextType } from '@/contexts/TimelineContext';
+import PostList from '@/features/posts/components/PostList/PostList';
 import { type PostWithProfile } from '@/types';
 
 type SearchResultClientProps = {
